refactor: re-organize directory tests
