+++ internal/provider/edgegw/edgegateway_resource.go
