+++ internal/tests/catalog/catalog_resource_test.go
@@ -1,10 +1,12 @@
-// Package tests provides the acceptance tests for the provider.
-package tests
+// Package catalog provides the acceptance tests for the provider.
+package catalog
 
 import (
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+
+	tests "github.com/orange-cloudavenue/terraform-provider-cloudavenue/internal/tests/common"
 )
 
 const testAccCatalogResourceConfig = `
@@ -20,8 +22,8 @@ func TestAccCatalogResource(t *testing.T) {
 	resourceName := "cloudavenue_catalog.test"
 
 	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { testAccPreCheck(t) },
-		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		PreCheck:                 func() { tests.TestAccPreCheck(t) },
+		ProtoV6ProviderFactories: tests.TestAccProtoV6ProviderFactories,
 		Steps: []resource.TestStep{
 			// Read testing
 			{
