+++ internal/tests/catalog/catalog_vapp_template_datasource_test.go
@@ -1,10 +1,12 @@
-// Package tests provides the acceptance tests for the provider.
-package tests
+// Package catalog provides the acceptance tests for the provider.
+package catalog
 
 import (
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+
+	tests "github.com/orange-cloudavenue/terraform-provider-cloudavenue/internal/tests/common"
 )
 
 //go:generate go run github.com/FrangipaneTeam/tf-doc-extractor@latest -filename $GOFILE -example-dir ../../examples -test
@@ -19,8 +21,8 @@ func TestAccCatalogVappTemplateDataSource(t *testing.T) {
 	dataSourceName := "data.cloudavenue_catalog_vapp_template.example"
 
 	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { testAccPreCheck(t) },
-		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		PreCheck:                 func() { tests.TestAccPreCheck(t) },
+		ProtoV6ProviderFactories: tests.TestAccProtoV6ProviderFactories,
 		Steps: []resource.TestStep{
 			// Read testing
 			{
