+++ internal/tests/common/acctest.go
@@ -15,14 +15,14 @@ import (
 // acceptance testing. The factory function will be invoked for every Terraform
 // CLI command executed to create a provider server to which the CLI can
 // reattach.
-var testAccProtoV6ProviderFactories = map[string]func() (tfprotov6.ProviderServer, error){
+var TestAccProtoV6ProviderFactories = map[string]func() (tfprotov6.ProviderServer, error){
 	"cloudavenue": providerserver.NewProtocol6WithError(provider.New("test")()),
 }
 
 // You can add code here to run prior to any test case execution, for example assertions
 // about the appropriate environment variables being set are common to see in a pre-check
 // function.
-func testAccPreCheck(t *testing.T) {
+func TestAccPreCheck(t *testing.T) {
 	if v := os.Getenv("CLOUDAVENUE_URL"); v == "" {
 		t.Fatal("CLOUDAVENUE_URL must be set for acceptance tests")
 	}
