+++ internal/tests/edgegw/edge_gateway_datasource_test.go
@@ -1,10 +1,12 @@
-// Package tests provides the acceptance tests for the provider.
-package tests
+// Package edgegw provides the acceptance tests for the provider.
+package edgegw
 
 import (
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+
+	tests "github.com/orange-cloudavenue/terraform-provider-cloudavenue/internal/tests/common"
 )
 
 const testAccEdgeGatewayDataSourceConfig = `
@@ -16,8 +18,8 @@ data "cloudavenue_edgegateway" "test" {
 func TestAccEdgeGatewayDataSource(t *testing.T) {
 	dataSourceName := "data.cloudavenue_edgegateway.test"
 	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { testAccPreCheck(t) },
-		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		PreCheck:                 func() { tests.TestAccPreCheck(t) },
+		ProtoV6ProviderFactories: tests.TestAccProtoV6ProviderFactories,
 		Steps: []resource.TestStep{
 			// Read testing
 			{
