+++ internal/tests/iam/iam_user_resource_test.go
@@ -1,25 +1,25 @@
-package tests
+// Package tests provides the acceptance tests for the provider.
+// Package iam provides the acceptance tests for the provider.
+package iam
 
 import (
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+
+	tests "github.com/orange-cloudavenue/terraform-provider-cloudavenue/internal/tests/common"
 )
 
-const testAccOrgUserDataSourceConfig = `
+const testAccOrgUserResourceConfig = `
 resource "cloudavenue_iam_user" "test" {
 	user_name   = "testuser"
 	description = "A test user"
 	role        = "Organization Administrator"
 	password    = "Th!s1sSecur3P@ssword"
-}
-
-data "cloudavenue_iam_user" "test" {
-	user_name = cloudavenue_iam_user.test.user_name
-}
+  }
 `
 
-const testAccOrgUserDataSourceConfigFull = `
+const testAccOrgUserResourceConfigFull = `
 resource "cloudavenue_iam_user" "test" {
 	user_name         = "testuserfull"
 	description       = "A test user"
@@ -32,23 +32,19 @@ resource "cloudavenue_iam_user" "test" {
 	take_ownership    = true
 	deployed_vm_quota = 10
 	stored_vm_quota   = 5
-}
-
-data "cloudavenue_iam_user" "test" {
-	user_name = cloudavenue_iam_user.test.user_name
-}
+  }
 `
 
-func TestAccOrgUserDataSource(t *testing.T) {
+func TestAccOrgUserResource(t *testing.T) {
 	resourceName := "cloudavenue_iam_user.test"
 
 	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { testAccPreCheck(t) },
-		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		PreCheck:                 func() { tests.TestAccPreCheck(t) },
+		ProtoV6ProviderFactories: tests.TestAccProtoV6ProviderFactories,
 		Steps: []resource.TestStep{
 			// Read testing
 			{
-				Config: testAccOrgUserDataSourceConfig,
+				Config: testAccOrgUserResourceConfig,
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestCheckResourceAttr(resourceName, "user_name", "testuser"),
 					resource.TestCheckResourceAttr(resourceName, "description", "A test user"),
@@ -57,7 +53,7 @@ func TestAccOrgUserDataSource(t *testing.T) {
 				),
 			},
 			{
-				Config: testAccOrgUserDataSourceConfigFull,
+				Config: testAccOrgUserResourceConfigFull,
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestCheckResourceAttr(resourceName, "user_name", "testuserfull"),
 					resource.TestCheckResourceAttr(resourceName, "description", "A test user"),
@@ -66,9 +62,19 @@ func TestAccOrgUserDataSource(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceName, "email", "foo@bar.com"),
 					resource.TestCheckResourceAttr(resourceName, "telephone", "1234567890"),
 					resource.TestCheckResourceAttr(resourceName, "full_name", "Test User"),
+					resource.TestCheckResourceAttr(resourceName, "take_ownership", "true"),
 					resource.TestCheckResourceAttr(resourceName, "deployed_vm_quota", "10"),
 					resource.TestCheckResourceAttr(resourceName, "stored_vm_quota", "5"),
 				),
+			},
+			// ImportState testing
+			{
+				ResourceName:      resourceName,
+				ImportState:       true,
+				ImportStateId:     "testuserfull",
+				ImportStateVerify: true,
+				// These fields can't be retrieved from user data
+				ImportStateVerifyIgnore: []string{"take_ownership", "password"},
 			},
 		},
 	})
