+++ internal/tests/vapp/vapp_datasource_test.go
@@ -1,11 +1,13 @@
-// Package tests provides the acceptance tests for the provider.
-package tests
+// Package vapp provides the acceptance tests for the provider.
+package vapp
 
 import (
 	"os"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+
+	tests "github.com/orange-cloudavenue/terraform-provider-cloudavenue/internal/tests/common"
 )
 
 const testAccVappDataSourceConfig = `
@@ -22,8 +24,8 @@ data "cloudavenue_vapp" "test" {
 func TestAccVappDataSource(t *testing.T) {
 	dataSourceName := "data.cloudavenue_vapp.test"
 	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { testAccPreCheck(t) },
-		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		PreCheck:                 func() { tests.TestAccPreCheck(t) },
+		ProtoV6ProviderFactories: tests.TestAccProtoV6ProviderFactories,
 		Steps: []resource.TestStep{
 			// Read testing.
 			{
